+++ crates/biome_js_analyze/tests/specs/suspicious/noRedeclare/invalid.ts
@@ -4,3 +4,5 @@ class C {
 		var a;
 	}
 }
+
+function f<T, T>() {}