+++ CHANGELOG.md
@@ -34,6 +34,26 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 - Code fixes of [useImportType](https://biomejs.dev/linter/rules/use-import-type) and [useExportType](https://biomejs.dev/linter/rules/use-export-type) now handle multiline statements ([#2041](https://github.com/biomejs/biome/issues/2041)).
   Contributed by @Conaclos
 
+- [noRedeclare](https://biomejs.dev/linter/rules/no-redeclare) no longer reports type parameter and parameter with identical names ([#1992](https://github.com/biomejs/biome/issues/1992)).
+
+  The following code is no longer reported:
+
+  ```ts
+  function f<a>(a: a) {}
+  ```
+
+  Contributed by @Conaclos
+
+- [noRedeclare](https://biomejs.dev/linter/rules/no-redeclare) now reports duplicate type parameters in a same declaration.
+
+  The following type parameters are now reported as a redeclaraion:
+
+  ```ts
+  function f<T, T>() {}
+  ```
+
+  Contributed by @Conaclos
+
 ### Parser